Create default.html
9Make a _layouts directory, and create file inside it called default.html. (Remember, you can make directories while making new files. See the main.css step if you forgot.)

This is our main layout that will contain repeated elements like our <head> and <footer>. Now we won't have to repeat that markup on every single page we create, making maintenance of our site much easier. So let's move those elements from index.html into default.html to get something that looks like this in the end. (image of code pasted in default.html)


Take note of the {{ page.title }} and {{ content }} tags in there. They're what Jekyll calls liquid tags, and these are used to inject content into the final web page. More on this in a bit.